+++ external/fv3fit/fv3fit/__init__.py
@@ -51,6 +51,6 @@ from .pytorch import (
 import fv3fit.train_microphysics
 import fv3fit.dataclasses
 import fv3fit.reservoir.train
-import fv3fit.reservoir.autoencoder
+import fv3fit.reservoir.transformers.autoencoder
 
 __version__ = "0.1.0"
