+++ external/fv3fit/fv3fit/reservoir/__init__.py
@@ -7,3 +7,5 @@ from .reservoir import Reservoir
 from .readout import ReservoirComputingReadout
 from .model import ReservoirComputingModel, HybridReservoirComputingModel
 from .domain import RankDivider
+from .transformers.autoencoder import Autoencoder
+from .transformers.sk_transformer import SkTransformer
